requirements.txt
python-dotenv
requests
tk